+++ src/adobe-analytics.ios.d.ts
@@ -1,5 +1,3 @@
-// use take the ios or android .d.ts files and copy/paste them here
-
 import { AdobeAnalyticsCommon } from './adobe-analytics.common';
 export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
     protected static _instance: AdobeAnalyticsCommon;
@@ -19,4 +17,10 @@ export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
         [key: string]: any;
     }): void;
     trackTimedActionEnd(action: string): void;
+    visitorAppendToURL(url: string): string;
+    trackLocation(location: CLLocation, additional: {
+        [key: string]: any;
+    }): void;
+    optIn(): void;
+    optOut(): void;
 }