+++ src/index.d.ts
@@ -1,8 +1,10 @@
+// use take the ios or android .d.ts files and copy/paste them here
 import { AdobeAnalyticsCommon } from './adobe-analytics.common';
+
 export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
     protected static _instance: AdobeAnalyticsCommon;
-    setContext(applicationContext: android.content.Context): void;
-    collectLifecycleData(activity: android.app.Activity, debugLogging?: boolean): void;
+    setContext(applicationContext: any): void;
+    collectLifecycleData(activity: any, debugLogging?: boolean): void;
     pauseCollectingLifecycleData(): void;
     trackState(state: string, additional: {
         [key: string]: any;
@@ -17,5 +19,8 @@ export declare class AdobeAnalytics extends AdobeAnalyticsCommon {
         [key: string]: any;
     }): void;
     trackTimedActionEnd(action: string): void;
-    private convertToHashMap(dictionary?);
+    visitorAppendToURL(url: string): string;
+    trackLocation(location: any, additional: { [key: string]: any; }): void;
+    optIn(): void;
+    optOut(): void;
 }