+++ changelog.md
@@ -1,2 +1,9 @@
+#1.5
+Added support for NativeScript 3.4
+
+Added support for privacy options (optIn and optOut)
+
+Added support for obtaining visitorId
+
 #1.4
 Added support for NativeScript 3.1